backend_django/requirements.txt
anyio==3.6.1
asgiref==3.5.2
backports.zoneinfo==0.2.1
bleach==5.0.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
click-log==0.4.0
colorama==0.4.5
deprecation==2.1.0
Django==4.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
docutils==0.19
dotty-dict==1.3.1
gitdb==4.0.9
GitPython==3.1.27
gotrue==0.5.4
h11==0.12.0
httpcore==0.14.7
httpx==0.21.3
idna==3.3
importlib-metadata==4.12.0
invoke==1.7.1
jaraco.classes==3.2.2
keyring==23.9.0
more-itertools==8.14.0
packaging==21.3
pkginfo==1.8.3
postgrest-py==0.10.2
pydantic==1.10.1
Pygments==2.13.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-decouple==3.6
python-gitlab==3.9.0
python-semantic-release==7.28.1
pytz==2022.2.1
pywin32-ctypes==0.2.0
readme-renderer==37.1
realtime==0.0.4
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sqlparse==0.4.2
storage3==0.3.4
supabase==0.5.8
tomlkit==0.10.2
tqdm==4.64.1
twine==3.8.0
typing-extensions==4.3.0
tzdata==2022.2
urllib3==1.26.12
webencodings==0.5.1
websockets==9.1
zipp==3.8.1
